Add core route/service/manager Add flow to get categories Add flow to get currencies Adjust more schemas from float fields to Decimal fields Adjust investment tests to parse Decimal to float

+++ schemas/account.py
@@ -1,5 +1,6 @@
 import datetime
 import uuid
+from decimal import Decimal
 
 from pydantic import BaseModel, Field, ConfigDict
 
@@ -65,9 +66,9 @@ class BalanceSchema(BaseModel):
 
     id: uuid.UUID = Field(..., serialization_alias='balanceEntryId', description='The id of the balance entry')
     account_id: uuid.UUID = Field(..., serialization_alias='accountId', description='Account identification')
-    previous_balance: float = Field(None, description="The balance from the past period")
-    incoming: float = Field(None, description="The amount o money that enter the account in the period")
-    outgoing: float = Field(None, description="The amount o money that leave the account in the period")
-    transactions: float = Field(None, description="The difference between incoming and outgoing")
-    earning: float = Field(None, description="How much the account profits in the period, if set")
-    balance: float = Field(None, description="How many money is in the account by the end of the period")
+    previous_balance: Decimal = Field(None, description="The balance from the past period")
+    incoming: Decimal = Field(None, description="The amount o money that enter the account in the period")
+    outgoing: Decimal = Field(None, description="The amount o money that leave the account in the period")
+    transactions: Decimal = Field(None, description="The difference between incoming and outgoing")
+    earning: Decimal = Field(None, description="How much the account profits in the period, if set")
+    balance: Decimal = Field(None, description="How many money is in the account by the end of the period")
